# tools/openapi.yaml
openapi: 3.0.3
info:
  version: 0.9.6
  title: LANraragi API
...
paths:
...
  /tankoubons:
    get:
      operationId: getTankoubonList
      x-mojo-to: api-tankoubon#get_tankoubon_list
      summary: Get all Tankoubons
      description: Get list of Tankoubons paginated.
      tags:
        - tankoubons
      parameters:
        - name: page
          in: query
          required: false
          description: >- 
            Page of the list of Tankoubons.  

            The amount of tanks per page depends on the server `archives_per_page` setting.
          schema:
            type: integer
      responses:
        '200':
          description: Tankoubons
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/TankoubonMetadataJson'
                  total:
                    type: integer
                    description: Total count of Tankoubons on the server
                  filtered:
                    type: integer
                    description: Number of Tankoubons on the current page
    put:
      security:
        - api_key: []
      operationId: createTankoubon
      x-mojo-to: api-tankoubon#create_tankoubon
      summary: 🔑 Create a Tankoubon
      description: Create a new Tankoubon or update the name of an existing one.
      tags:
        - tankoubons
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Name of the Tankoubon
                tankid:
                  type: string
                  description: ID of an existing Tankoubon, if you want to change its name.
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Name of the Tankoubon
                tankid:
                  type: string
                  description: ID of an existing Tankoubon, if you want to change its name.
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum: 
                      - create_tankoubon
                  tankoubon_id:
                    type: string
                    description: ID of the created/modified Tankoubon
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /tankoubons/{id}:
    parameters:
        - name: id
          in: path
          required: true
          description: ID of the Tankoubon desired
          schema:
            type: string
    get:
      operationId: getTankoubon
      x-mojo-to: api-tankoubon#get_tankoubon
      summary: Get a single Tankoubon
      description: >-
        Get the details of the specified tankoubon ID, with the archives list paginated.  

        The amount of archives per page depends on the server `archives_per_page` setting.
      tags:
        - tankoubons
      parameters:
        - name: include_full_data
          in: query
          required: false
          description: If set to true, a `full_data` array will be added to the response with full Archive metadata.
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          description: Page of the Archives list. You can use -1 to return all archives if needed.  
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Tankoubon
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      archives:
                        type: array
                        description: Paginated list of archive IDs or tankoubon IDs.
                        items:
                          type: string
                          oneOf:
                            - minLength: 40
                              maxLength: 40
                              pattern: "^[0-9a-f]{40}$"
                            - minLength: 15
                              maxLength: 15
                              pattern: "^TANK_[0-9]{10}$"
                        uniqueItems: true
                      full_data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ArchiveMetadataJson'
                  total:
                    type: integer
                    description: The total amount of archives in this Tankoubon.
                  filtered:
                    type: integer
                    description: The amount of archives in the current page.
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
    put:
      security:
        - api_key: []
      operationId: updateTankoubon
      x-mojo-to: api-tankoubon#update_tankoubon
      summary: 🔑 Update Tankoubon metadata/contents
      description: >- 
        Modify the full metadata (name, summary, additional tags) or the contents of a Tankoubon.  

        If you only need to change the name of Tank, consider just using PUT `/api/tankoubons`.  


        Note: If there is no need to update something in one of the `data` keys, do not send the key -- This otherwise can result in unwanted results.
      tags:
        - tankoubons
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Tankoubon to update.
          schema:
            type: string
      requestBody:
        required: true
        description: Json with 2 optional keys "archives" and "metadata"
        content:
          application/json:
            schema:
              type: object
              properties:
                archives:
                  description: Ordered array with the IDs of the archives.
                  type: array
                  items:
                    type: string
                    minLength: 40
                    maxLength: 40
                metadata:
                  description: Metadata parameters -- All are optional.
                  type: object
                  properties:
                    name:
                      type: string
                      description: Name of the Tankoubon
                    summary:
                      type: string
                      description: Summary of the Tankoubon
                    tags:
                      type: string
                      description: Additional tags for the Tankoubon, in LRR comma-separated format.
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
    delete:
      security:
        - api_key: []
      operationId: deleteTankoubon
      x-mojo-to: api-tankoubon#delete_tankoubon
      summary: 🔑 Delete Tankoubon
      description: Remove a Tankoubon from the server. This doesn't delete underlying Archives.
      tags:
        - tankoubons
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /tankoubons/{id}/thumbnail:
    parameters:
        - name: id
          in: path
          required: true
          description: ID of the Tankoubon desired
          schema:
            type: string
    get:
      operationId: getTankoubonThumbnail
      x-mojo-to: api-tankoubon#serve_tankoubon_thumbnail
      summary: Get Tankoubon thumbnail
      description: >-
        Get the cover thumbnail for a given Tankoubon. 
        By default, the thumbnail is sourced from the first page of the first archive. 
        This endpoint will return a placeholder image if it doesn't already exist. 
        If you want to queue generation of the thumbnail in the background, 
        you can use the no_fallback query parameter. 
      tags:
        - tankoubons
      parameters:
        - name: no_fallback
          in: query
          required: false
          description: >-
            Disables the placeholder image, queues the thumbnail for extraction
            and returns a JSON with code 202. This parameter does nothing if the
            image already exists. (You will get the image with code 200 no
            matter what)
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
      responses:
        '200':
          description: If the thumbnail was already extracted, you get it directly.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '202':
          description: The thumbnail is queued for extraction. Use `/api/minion/:jobid` to track when your thumbnail is ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MinionJobResponse'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
    put:
      security:
        - api_key: []
      operationId: updateTankoubonThumbnail
      x-mojo-to: api-tankoubon#update_tankoubon_thumbnail
      summary: 🔑 Update Tankoubon thumbnail
      description: >-
        Set the cover thumbnail for the given Tankoubon using a global page
        number that spans all archives in the tank (in order). The global page
        is translated to the correct archive and local page automatically.
      tags:
        - tankoubons
      parameters:
        - name: page
          in: query
          required: true
          description: >-
            Global 1-indexed page number across all archives in the tankoubon.
            Page 1 is the first page of the first archive, and so on.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Tankoubon thumbnail updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum:
                      - update_tankoubon_thumbnail
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                  new_thumbnail:
                    type: string
                    description: Path to the new thumbnail file
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /tankoubons/{id}/progress/{page}:
    put:
      operationId: updateTankoubonProgress
      x-mojo-to: api-tankoubon#update_tank_progress
      summary: Update Tankoubon reading progression
      description: >-
        Tell the server which page of this Tankoubon you're currently
        reading, so that it updates its internal reading progression accordingly.

        The page number is global across all Archives in the tank 
        (If a tank has two Archives with 20 and 25 pages, page 26 will be page 6 in Archive #2).

        ⚠ If the server is configured to use clientside progress tracking, this
        API call will return an error!

        Make sure to check using `/api/info` whether the server tracks reading
        progression or not before calling this endpoint.
      tags:
        - tankoubons
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Tankoubon to update.
          schema:
            type: string
            minLength: 15
            maxLength: 15
        - name: page
          in: path
          required: true
          description: >-
            Global 1-indexed page number to update the reading progress to.
            Must be a positive integer.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 15
                    maxLength: 15
                    description: ID of the tankoubon updated
                  operation:
                    type: string
                    description: Name of the operation
                    enum:
                      - update_tank_progress
                  page:
                    type: integer
                    description: Updated reading progress
                  lastreadtime:
                    type: integer
                    description: Last read time in epoch seconds
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                    description: Returns 1 if successful, else 0
              example:
                id: TANK_1589141306
                operation: update_tank_progress
                page: 5
                lastreadtime: 123943543
                success: 1
        '400':
          description: Generic error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                operation: update_tank_progress
                error: Server-side Progress Tracking is disabled on this instance.
                success: 0
        '401':
          description: Authentication required when authprogress is enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                operation: update_tank_progress
                error: This operation requires authentication.
                success: 0
  /tankoubons/{id}/{archive}:
    put:
      security:
        - api_key: []
      operationId: addToTankoubon
      x-mojo-to: api-tankoubon#add_to_tankoubon
      summary: 🔑 Add an Archive to a Tankoubon
      description: Append an archive at the final position of a Tankoubon.
      tags:
        - tankoubons
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Tankoubon to update.
          schema:
            type: string
        - name: archive
          in: path
          required: true
          description: ID of the Archive to append.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
    delete:
      security:
        - api_key: []
      operationId: removeFromTankoubon
      x-mojo-to: api-tankoubon#remove_from_tankoubon
      summary: 🔑 Remove an Archive from a Tankoubon
      tags:
        - tankoubons
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the Tankoubon to update.
          schema:
            type: string
        - name: archive
          in: path
          required: true
          description: ID of the Archive to remove.
          schema:
            type: string
            minLength: 40
            maxLength: 40
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header
      description: 'Use Authorization: Bearer <base64(api_key)>'
  schemas:
...
    OperationResponse:
      type: object
      required:
        - operation
        - success
      properties:
        operation:
          type: string
          description: Name of operation
        error:
          type: string
          description: Error message if any
        successMessage:
          type: string
          description: Success message if any
        success:
          type: integer
          description: Returns 1 if operation was successful, else 0
          enum:
            - 0
            - 1
    MinionJobResponse:
      type: object
      description: JSON object for a queued Minion job.
      required:
        - job
        - operation
        - success
      properties:
        job:
          type: integer
          description: Minion job ID
        operation:
          type: string
          description: Name of operation
        success:
          type: integer
          description: Returns 1 if operation was successful, else 0
          enum: 
            - 1
            - 0
    ArchiveMetadataJson:
      type: object
      description: JSON object for archive metadata for most API responses
      required:
        - arcid
        - title
        - filename
        - tags
        - isnew
        - extension
        - progress
        - pagecount
        - lastreadtime
        - size
      properties:
        arcid:
          type: string
          description: Unique identifier for the archive (40-char SHA1) or tankoubon (15-char TANK_* ID).
          pattern: "^([0-9a-f]{40}|TANK_[0-9]{10})$"
        title:
          type: string
          description: Title of the archive
        filename:
          type: string
          description: Filename of the archive
        tags:
          type: string
          description: Comma-separated list of tags associated with the archive
        summary:
          type: string
          nullable: true
          description: Summary description of the archive
        isnew:
          type: string
          enum:
            - "true"
            - "false"
          description: Whether the archive is newly added.
        extension:
          type: string
          description: File extension of the archive
        progress:
          type: integer
          description: Reading progress (page number)
        pagecount:
          type: integer
          description: Total number of pages in the archive
        lastreadtime:
          type: integer
          description: Unix timestamp of when the archive was last read
        size:
          type: integer
          description: Size of the archive in bytes
        toc:
          type: array
          description: Table of contents for the archive
      example:
        arcid: 28697b96f0ac5858be2614ed10ca47742c9522fd
        isnew: "false"
        extension: rar
        pagecount: 34
        progress: 3
        tags: parody:fate grand order,  group:wadamemo,  artist:wada rco,  artbook,  full color
        lastreadtime: 1337038281
        title: Fate GO MEMO
        filename: Fate GO MEMO
        summary: null
        size: 1234567
        toc: 
          - name: Chapter 1
            page: 1
          - name: Chapter 2
            page: 13
          - name: Chapter 3
            page: 25
...
    TankoubonMetadataJson:
      type: object
      description: Tankoubon metadata
      required:
        - tankid
      properties:
        archives:
          type: array
          description: Array of archive IDs
          items:
            type: string
            minLength: 40
            maxLength: 40
        tankid:
          type: string
          description: ID of the tankoubon
        name:
          type: string
          description: Name of the tankoubon
      example:
        tankid: TANK_1749925516
        name: sailor moon
        archives:
          - e69e43e1355267f7d32a4f9b7f2fe108d2401ebf
          - e69e43e1355267f7d32a4f9b7f2fe108d2401ebg
...
servers:
  - url: https://lrr.tvc-16.science/api